+++ react/helpers/index.tsx
@@ -1,9 +1,19 @@
 import React, { Fragment } from 'react'
+import ReactMarkdown from 'react-markdown'
 
-export const convertBreakToBrHTML = (text: string, keyPrefix: string) =>
-  text.split(/\n/g).map((paragraph, index) => (
-    <Fragment key={`${keyPrefix}-${index}`}>
-      {paragraph}
-      <br />
-    </Fragment>
-  ))
+export const convertBreakToBrHTML = (text: string, keyPrefix: string) => {
+  const textSplitted = text?.split(/\n/g)
+
+  return textSplitted?.map((paragraph, index) =>
+    paragraph?.trim() ? (
+      <ReactMarkdown
+        components={{
+          p: Fragment,
+        }}
+        key={`${keyPrefix}-${index}`}
+      >
+        {paragraph}
+      </ReactMarkdown>
+    ) : undefined
+  )
+}