+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Seller description limiter.
+
 ## [0.0.9] - 2023-07-17
 
 ## [0.0.8] - 2023-07-17